# API/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the documentation for the golang api of appolo-api",
        "title": "Appolo API",
        "contact": {
            "name": "API Support",
            "email": "pierre.saintsorny@gmail.com"
        },
        "version": "1"
    },
    "basePath": "/1.0",
    "paths": {
...
        "/binance/order/buy": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Buy from binance using market order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "my_first_strategy",
                        "name": "strategy",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "0.01",
                        "name": "quantity",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/binance/order/sell": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "sell from binance using market order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "my_first_strategy",
                        "name": "strategy",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "0.01",
                        "name": "quantity",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/binance/price": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get price for a given symbol",
                "parameters": [
                    {
                        "type": "string",
                        "description": "BTCUSDT",
                        "name": "symbol",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Prices"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
...
        "/strategy": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get one or multiple strategy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "my_first_strategy",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Strategy"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new strategy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "BTCUSDT",
                        "name": "symbol",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "my_first_strategy_BTCUSDT",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "1h",
                        "name": "interval",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "10",
                        "name": "leverage",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "undefined",
                        "name": "account_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a strategy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "my_first_strategy",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        },
        "/strategy/result": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get the total pnl and commission for a strategy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "my_strategy",
                        "name": "strategy_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TotalPNLAndCommission"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "models.Order": {
            "type": "object",
            "properties": {
                "executedQty": {
                    "type": "string"
                },
                "fills": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "commission": {
                                "type": "string"
                            },
                            "price": {
                                "type": "string"
                            },
                            "qty": {
                                "type": "string"
                            }
                        }
                    }
                },
                "orderId": {
                    "type": "number"
                },
                "side": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "transactTime": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.Prices": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "models.Strategy": {
            "type": "object",
            "properties": {
                "accountName": {
                    "type": "string"
                },
                "currencyName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "interval": {
                    "type": "string"
                },
                "leverage": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.TotalPNLAndCommission": {
            "type": "object",
            "properties": {
                "commission": {
                    "type": "number"
                },
                "pnl": {
                    "type": "number"
                }
            }
        },
...
    }
}